requirements.txt
requests
pandas
pypdf
tqdm
python-dotenv

langgraph
langchain
langchain-community
langchain-core
langchain-openai

semantic_router
pinecone
serpapi
google-search-results